Allow down casting query trail interface and union types
This is part of fixing
https://github.com/davidpdrsn/juniper-eager-loading/issues/15.

In order to perform eager loading on a `Vec<User>` you need to know
which associations are being queried. The `QueryTrail<User, _>` type has
that information.

However, this doesn't work for interfaces since if you have an interface
called for example `Entity` you will not get a `QueryTrail<User, _>` but
instead a `QueryTrail<Entity, _>`. That doesn't work to eager load
users, which means you cannot eager at all. This is because `Entity` is
an enum with a variant for each type that implements that interface.

With this change we also generate implementations of `Into` for
`QueryTrail<Entity, _>` into each type that implements it. That allows
you get a `QueryTrail<User, _>` to do eager loading with.