Overhaul directive parsing
In order to support `DateTime` scalars with or without time zones I
would like have a schema directive to customize it. However the previous
directive parsing code was not easy to extend to support new types.
Especially supporting the same directives across types wasn't easy.

The idea behind this patch is to introduce a trait called
`FromDirective` that defines how to parse a `Directive` into some type.
Kinda like serde's `Deserialize`.

With that trait we can take a list of directives and for example parse
it into the directives supported by fields, enums, or even scalars.

I have also overhauled the related error messages so they should now be
more clear.