Remove use of `juniper::graphql_object!`
Fixes #55

This isn't complete yet. This patch currently removes support for having
descriptions on field arguments. That is because getting those to work
[requires funky stuff][funky] in the juniper proc macro invocation.

While it could totally be made to work I would rather wait since it
requires a bit of refactoring and since it should work seamlessly once
[RFC 2564](https://github.com/rust-lang/rust/issues/60406) is stable,
which seems to happen in 1.39.

[funky]: https://docs.rs/juniper_codegen/0.13.2/juniper_codegen/attr.object.html#customization-documentation-renaming-